Move numpyro single mcmc to sampler object
