+++ tests/superphot_plus/samplers/test_numpyro_sampler.py
@@ -1,21 +1,21 @@
-import os
-
 import numpy as np
 import pytest
 
-from superphot_plus.fit_numpyro import numpyro_single_file, run_mcmc, trunc_norm
-from superphot_plus.lightcurve import Lightcurve
+from superphot_plus.priors.fitting_priors import PriorFields
+from superphot_plus.samplers.numpyro_sampler import NumpyroSampler, run_mcmc, trunc_norm
 
 
 def test_trunc_norm(jax_key):
     """Test that the trunc_norm() function, which returns numpyro's
     TruncatedNormal distribution, works as expected.
     """
-    unit_untrunc = trunc_norm(loc=0.0, scale=1.0, low=None, high=None)  # should default to a Normal dist
+    fields = PriorFields(None, None, 0, 1.0)
+    unit_untrunc = trunc_norm(fields)  # should default to a Normal dist
     assert unit_untrunc.mean == 0.0
     assert unit_untrunc.variance == 1.0
 
-    unit_trunc = trunc_norm(loc=0.0, scale=1.0, low=-0.5, high=0.5)
+    fields = PriorFields(-0.5, 0.5, 0, 1.0)
+    unit_trunc = trunc_norm(fields)
 
     assert np.all(unit_trunc.log_prob(np.array([-0.4, 0.0, 0.4])) > -np.inf)
     assert np.all(unit_trunc.log_prob(np.array([-0.6, 0.6, 10.0])) == -np.inf)
@@ -23,38 +23,35 @@ def test_trunc_norm(jax_key):
     assert np.all(np.abs(unit_trunc.sample(jax_key, sample_shape=(10, 10))) < 0.5)
 
 
-def test_mcmc_missing_band(single_ztf_lightcurve_compressed):
+def test_mcmc_missing_band(single_ztf_lightcurve_object):
     """Test that run_mcmc exists out with missing band data."""
-    lc = Lightcurve.from_file(single_ztf_lightcurve_compressed)
-    lc.filter_by_band(
+    single_ztf_lightcurve_object.filter_by_band(
         [
             "r",
         ]
     )
-    assert run_mcmc(lc) is None
+    assert run_mcmc(single_ztf_lightcurve_object) is None
 
 
-def test_nonimplemented_sampler(tmp_path, single_ztf_lightcurve_compressed):
+def test_nonimplemented_sampler(single_ztf_lightcurve_object, ztf_priors):
     """Tests that run_mcmc exists out when non-implemented sampler
     name is given.
     """
     with pytest.raises(ValueError):
-        numpyro_single_file(single_ztf_lightcurve_compressed, tmp_path, sampler="sampler")
+        sampler = NumpyroSampler()
+        _ = sampler.run_single_curve(single_ztf_lightcurve_object, priors=ztf_priors, sampler="sampler")
 
 
-def test_numpyro_nuts(tmp_path, single_ztf_lightcurve_compressed):
+def test_numpyro_nuts(ztf_priors, single_ztf_lightcurve_object):
     """Test that we generated a new file, that its samples that can be
     read, and the mean of samples generated is within a certain range of
     expected values."""
-    numpyro_single_file(single_ztf_lightcurve_compressed, tmp_path, sampler="NUTS")
-
-    # Check ouput existence
-    output_file = os.path.join(tmp_path, "ZTF22abvdwik_eqwt_NUTS.npz")
-    assert os.path.exists(output_file)
-
+    sampler = NumpyroSampler()
+    posterior_samples = sampler.run_single_curve(
+        single_ztf_lightcurve_object, priors=ztf_priors, sampler="NUTS"
+    )
     # Check output length
-    params = np.load(output_file)["arr_0"]
-    assert len(params) == 300
+    assert len(posterior_samples.samples) == 300
 
     # Check output values
     expected = [
@@ -73,23 +70,20 @@ def test_numpyro_nuts(tmp_path, single_ztf_lightcurve_compressed):
         5.76787619e-01,
         8.59146651e-01,
     ]
-    sample_mean = np.mean(params, axis=0)
+    sample_mean = posterior_samples.sample_mean()
     assert np.all(np.isclose(sample_mean, expected, rtol=0.5))
 
 
-def test_numpyro_svi(tmp_path, single_ztf_lightcurve_compressed):
+def test_numpyro_svi(ztf_priors, single_ztf_lightcurve_object):
     """Test that we generated a new file, that its samples that can be
     read, and the mean of samples generated is within a certain range of
     expected values."""
-    numpyro_single_file(single_ztf_lightcurve_compressed, tmp_path, sampler="svi")
-
-    # Check ouput existence
-    output_file = os.path.join(tmp_path, "ZTF22abvdwik_eqwt_svi.npz")
-    assert os.path.exists(output_file)
-
+    sampler = NumpyroSampler()
+    posterior_samples = sampler.run_single_curve(
+        single_ztf_lightcurve_object, priors=ztf_priors, sampler="svi"
+    )
     # Check output length
-    params = np.load(output_file)["arr_0"]
-    assert len(params) == 100
+    assert len(posterior_samples.samples) == 100
 
     # Check output values
     expected = [
@@ -108,5 +102,5 @@ def test_numpyro_svi(tmp_path, single_ztf_lightcurve_compressed):
         5.72e-01,
         8.57e-01,
     ]
-    sample_mean = np.mean(params, axis=0)
+    sample_mean = posterior_samples.sample_mean()
     assert np.all(np.isclose(sample_mean, expected, rtol=0.5))
